+++ src/hera/task.py
@@ -691,7 +691,7 @@ class Task:
                 value = json.dumps(value)
             if name in param_name_cache:
                 continue  # user override of a kwarg
-            parameters.append(IoArgoprojWorkflowV1alpha1Parameter(name=name, default=value))
+            parameters.append(IoArgoprojWorkflowV1alpha1Parameter(name=name, default=value, value=value))
         return parameters
 
     def get_param_script_portion(self) -> str:
@@ -966,10 +966,8 @@ class Task:
         if self.tolerations is None:
             return []
 
-        ts = []
-        for t in self.tolerations:
-            ts.append(ArgoToleration(key=t.key, effect=t.effect, operator=t.operator, value=t.value))
-        return ts if ts else []
+        ts = [t.to_argo_toleration() for t in self.tolerations]
+        return ts
 
     def get_continue_on(self) -> Optional[IoArgoprojWorkflowV1alpha1ContinueOn]:
         """Assembles and returns the `continue_on` task setting"""